+++ CHANGELOG.md
@@ -24,6 +24,7 @@ ambiguous situations.
 
 ### Fixed
 - Forum crawling crashing when parsing empty (= 0 messages) threads
+- Forum crawling crashing when a forum has no threads at all
 
 ## 3.4.1 - 2022-08-17